Fix enable/disable

+++ src/core/localdatabase/localdatabasemanager.cpp
@@ -10,6 +10,7 @@
 #include "localmessagedatabase.h"
 #include "localmessagelogger.h"
 #include "localroomsdatabase.h"
+#include "ruqolaglobalconfig.h"
 
 LocalDatabaseManager::LocalDatabaseManager()
     : mMessageLogger(std::make_unique<LocalMessageLogger>())
@@ -25,7 +26,9 @@ LocalDatabaseManager::~LocalDatabaseManager() = default;
 void LocalDatabaseManager::addMessage(const QString &accountName, const QString &roomName, const Message &m)
 {
     mMessageLogger->addMessage(accountName, roomName, m);
-    mMessagesDatabase->addMessage(accountName, roomName, m);
+    if (RuqolaGlobalConfig::self()->storeMessageInDataBase()) {
+        mMessagesDatabase->addMessage(accountName, roomName, m);
+    }
     // Update timestamp.
     // mGlobalDatabase->updateTimeStamp(accountName, roomName, m.timeStamp()); ???
 }
@@ -33,36 +36,52 @@ void LocalDatabaseManager::addMessage(const QString &accountName, const QString 
 void LocalDatabaseManager::deleteMessage(const QString &accountName, const QString &roomName, const QString &messageId)
 {
     mMessageLogger->deleteMessage(accountName, roomName, messageId);
-    mMessagesDatabase->deleteMessage(accountName, roomName, messageId);
+    if (RuqolaGlobalConfig::self()->storeMessageInDataBase()) {
+        mMessagesDatabase->deleteMessage(accountName, roomName, messageId);
+    }
 }
 
 QVector<Message>
 LocalDatabaseManager::loadMessages(const QString &accountName, const QString &roomName, qint64 startId, qint64 endId, qint64 numberElements) const
 {
-    return mMessagesDatabase->loadMessages(accountName, roomName, startId, endId, numberElements);
+    if (RuqolaGlobalConfig::self()->storeMessageInDataBase()) {
+        return mMessagesDatabase->loadMessages(accountName, roomName, startId, endId, numberElements);
+    }
+    return {};
 }
 
 void LocalDatabaseManager::addRoom(const QString &accountName, Room *room)
 {
-    mRoomsDatabase->addRoom(accountName, room);
+    if (RuqolaGlobalConfig::self()->storeMessageInDataBase()) {
+        mRoomsDatabase->addRoom(accountName, room);
+    }
 }
 
 void LocalDatabaseManager::deleteRoom(const QString &accountName, const QString &roomId)
 {
-    mRoomsDatabase->deleteRoom(accountName, roomId);
+    if (RuqolaGlobalConfig::self()->storeMessageInDataBase()) {
+        mRoomsDatabase->deleteRoom(accountName, roomId);
+    }
 }
 
 void LocalDatabaseManager::updateTimeStamp(const QString &accountName, const QString &roomName, qint64 timestamp, LocalGlobalDatabase::TimeStampType type)
 {
-    mGlobalDatabase->updateTimeStamp(accountName, roomName, timestamp, type);
+    if (RuqolaGlobalConfig::self()->storeMessageInDataBase()) {
+        mGlobalDatabase->updateTimeStamp(accountName, roomName, timestamp, type);
+    }
 }
 
 void LocalDatabaseManager::removeTimeStamp(const QString &accountName, const QString &roomName, LocalGlobalDatabase::TimeStampType type)
 {
-    mGlobalDatabase->removeTimeStamp(accountName, roomName, type);
+    if (RuqolaGlobalConfig::self()->storeMessageInDataBase()) {
+        mGlobalDatabase->removeTimeStamp(accountName, roomName, type);
+    }
 }
 
 qint64 LocalDatabaseManager::timeStamp(const QString &accountName, const QString &roomName, LocalGlobalDatabase::TimeStampType type)
 {
-    return mGlobalDatabase->timeStamp(accountName, roomName, type);
+    if (RuqolaGlobalConfig::self()->storeMessageInDataBase()) {
+        return mGlobalDatabase->timeStamp(accountName, roomName, type);
+    }
+    return -1;
 }
